debugging train

+++ generate_sentence_embedding_file.py
@@ -124,7 +124,7 @@ def split_data(config):
     del text_len_test
     del combine
     # print(trainn[0][6],trainn[1][6])
-    print(trainn[:, 6].astype(np.int64))
+    print(trainn[:, 6].astype(np.float32))
     trainn = trainn[trainn[:, 6].astype(np.int64).argsort()]
     # for sample in trainn:
     #     print(sample[6])
